pyproject.toml
[tool.poetry]
name = "spotcli"
version = "1.1.2"
description = "CLI for Spot Elastigroups management"
authors = ["Dmitry Kovalenko <dmytro.kovalenko@ironsrc.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
importlib-metadata = {version = "^1.0", python = "<3.8"}
rich = "^9.0.1"
click = "^7.1.2"
durations = "^0.3.3"
spotinst-sdk = "^1.0.51"
python-consul2 = "^0.1.5"
stringcase = "^1.2.0"
attrs = "^20.2.0"
python-configuration = {extras = ["yaml"], version = "^0.8.1"}
semver = "^2.13.0"
requests = "^2.24.0"
boto3 = "^1.16.17"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
black = { version = "^20.8b1", allow-prereleases = true }
flake8 = "^3.8.4"
isort = "^5.6.4"

[tool.poetry.scripts]
spotcli = "spotcli.cli:main"

[tool.pytest.ini_options]
python_files = "test_*"
python_classes = "*Tests"
python_functions = "test_*"
testpaths = ["test"]
markers = ["elastigroup: Tests for Elastigroup module", "unit: Unit tests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
